refactor: split E2E tests into separate files and improve test performance
