+++ docs/guides/diagrams/kuadrant-integration-arch.drawio
