+++ docs/guides/diagrams/kuadrant-integration-arch.png
